Extract `cvd/fetch/de_android_sparse.{cc,h}`
Rename `DeAndroidSparse` to `DeAndroidSparse2` to avoid conflicts with
image_aggregator.

Bug: b/424021019
Test: bazel test '//...' --test_summary=terse
